Bug fix in SLAMSWLQ; fix type mismatches.

+++ SRC/zhetri2x.f
@@ -137,7 +137,7 @@
 *  =====================================================================
 *
 *     .. Parameters ..
-      REAL               ONE
+      DOUBLE PRECISION   ONE
       COMPLEX*16            CONE, ZERO
       PARAMETER          ( ONE = 1.0D+0,
      $                   CONE = ( 1.0D+0, 0.0D+0 ),
